agent/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "lucid-it-agent"
version = "0.1.0"
description = "AI-powered IT helpdesk automation agent"
requires-python = ">=3.10"
license = {text = "Apache-2.0"}
authors = [
    { name = "Alton Lord" }
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "griptape[drivers-prompt-ollama]>=1.8.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    "httpx>=0.27.0",
    "tenacity>=8.2.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
# Cloud LLM providers
openai = ["griptape[drivers-prompt-openai]>=1.8.0"]
anthropic = ["griptape[drivers-prompt-anthropic]>=1.8.0"]
all-llms = [
    "griptape[drivers-prompt-openai,drivers-prompt-anthropic]>=1.8.0"
]

# Development dependencies
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
    "pre-commit>=3.0",
]

[project.scripts]
lucid-agent = "agent.runtime.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/agent"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]  # Line length handled by black

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "integration: marks tests as integration tests (may require external services)",
    "manual: marks tests that should only be run manually",
]
filterwarnings = [
    "ignore::DeprecationWarning",
]
